+++ services/hostsService.js
@@ -13,7 +13,7 @@ export async function getAllHosts(req, res, next) {
         ? {
             name: {
               contains: name, // Perform a case-insensitive match (substring search)
-              mode: 'insensitive',
+              mode: "insensitive",
             },
           }
         : {}, // If no name query parameter, return all hosts
@@ -27,9 +27,20 @@ export async function getAllHosts(req, res, next) {
 
 export async function createHost(req, res, next) {
   try {
+    const { username, email, password } = req.body;
+
+    // Check if required fields are provided
+    if (!username || !email || !password) {
+      return res
+        .status(400)
+        .json({ error: "Username, email, and password are required" });
+    }
+
+    // Create the host
     const host = await prisma.host.create({
       data: req.body,
     });
+
     res.status(201).json(host);
   } catch (error) {
     next(error);
@@ -50,11 +61,22 @@ export async function getHostById(req, res, next) {
 
 export async function updateHost(req, res, next) {
   try {
-    const host = await prisma.host.update({
+    // Check if the host exists
+    const existingHost = await prisma.host.findUnique({
+      where: { id: req.params.id },
+    });
+
+    if (!existingHost) {
+      return res.status(404).json({ error: "Host not found" });
+    }
+
+    // Update the host
+    const updatedHost = await prisma.host.update({
       where: { id: req.params.id },
       data: req.body,
     });
-    res.status(200).json(host);
+
+    res.status(200).json(updatedHost);
   } catch (error) {
     next(error);
   }
@@ -62,16 +84,26 @@ export async function updateHost(req, res, next) {
 
 export async function deleteHost(req, res, next) {
   try {
-    await prisma.property.updateMany({
-      where: { hostId: req.params.id },
-      data: { hostId: null }, // Set hostId to null in related properties
-    });
-    const deletedHost = await prisma.host.delete({
+    // Check if the host exists
+    const existingHost = await prisma.host.findUnique({
       where: { id: req.params.id },
     });
-    if (!deletedHost) {
-      throw new NotFoundError("host", req.params.id);
+
+    if (!existingHost) {
+      return res.status(404).json({ error: "Host not found" });
     }
+
+    // Set hostId to null in related properties
+    await prisma.property.updateMany({
+      where: { hostId: req.params.id },
+      data: { hostId: null },
+    });
+
+    // Delete the host
+    await prisma.host.delete({
+      where: { id: req.params.id },
+    });
+
     res.status(200).json({ message: "Host deleted" });
   } catch (error) {
     next(error);
